// src/sdks/systemV2ApiDocs.ts
export type getSystemDictenumPageProps = {
  dictLabel?: string | null | undefined;
  dictType: string;
  dictValue?: string | null | undefined;
  pageNum: number;
  pageSize: number;
  sourceType?: string | null | undefined;
};
// ...
export interface DictTypeResp {
  dictName?: string;

  dictType?: string;

  id?: number;
}